MB-30044: Make monotonic.h templates use explicit constructors
Also make a couple of refactors to clang-tidy the rest of the file.

Reviewed-on: http://review.couchbase.org/c/kv_engine/+/132236
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/dcp/producer.cc
@@ -103,7 +103,7 @@ void DcpProducer::BufferLog::unpauseIfSpaceAvailable() {
                 "DcpProducer::BufferLog is full; ackedBytes:{}"
                 ", bytesSent:{}, maxBytes:{}",
                 producer.logHeader(),
-                uint64_t(ackedBytes),
+                ackedBytes,
                 uint64_t(bytesOutstanding),
                 uint64_t(maxBytes));
     } else {
@@ -124,7 +124,7 @@ void DcpProducer::BufferLog::acknowledge(size_t bytes) {
                     "DcpProducer::BufferLog is no longer full; ackedBytes:{}"
                     ", bytesSent:{}, maxBytes:{}",
                     producer.logHeader(),
-                    uint64_t(ackedBytes),
+                    ackedBytes,
                     uint64_t(bytesOutstanding),
                     uint64_t(maxBytes));
             producer.scheduleNotify();
